Load user into the website
The website now loads and stores user in its cookies. This makes sure that the user has a different navigation bar than that of a site visitor. Authenticated user has a profile icon on the navigation bar instead of the "Login" button where they can access their account settings, orders or log out of the website.